Character movement V0.1
Added a dash. It can be used  to the left or right by pressing space while moving on those directions. It has a cooldown before you can use it again of 0.5s and lasts 0.1s. Can be performed in the air, in which case, the player doesn't fall until the dash stops.